+++ src/main/java/net/mrscauthd/boss_tools/capability/IEnergyStorageDelegateHolder.java
@@ -1,4 +1,4 @@
-package net.mrscauthd.boss_tools.machines.machinetileentities;
+package net.mrscauthd.boss_tools.capability;
 
 import net.minecraftforge.energy.IEnergyStorage;
