+++ src/main/java/net/mrscauthd/boss_tools/capability/IOxygenStorage.java
@@ -1,4 +1,4 @@
-package net.mrscauthd.boss_tools.armor.oxygensystem;
+package net.mrscauthd.boss_tools.capability;
 
 public interface IOxygenStorage {
