+++ src/main/java/net/mrscauthd/boss_tools/capability/IOxygenStorageHolder.java
@@ -1,4 +1,4 @@
-package net.mrscauthd.boss_tools.armor.oxygensystem;
+package net.mrscauthd.boss_tools.capability;
 
 public interface IOxygenStorageHolder {
 	void onOxygenChanged(IOxygenStorage oxygenStorage, int oxygenDelta);
