+++ src/main/java/net/mrscauthd/boss_tools/gui/CompressorGuiGui.java
@@ -32,7 +32,7 @@ import net.minecraftforge.items.ItemStackHandler;
 import net.minecraftforge.items.SlotItemHandler;
 import net.mrscauthd.boss_tools.BossToolsMod;
 import net.mrscauthd.boss_tools.BossToolsModElements;
-import net.mrscauthd.boss_tools.machines.machinetileentities.ItemStackToItemStackTileEntity;
+import net.mrscauthd.boss_tools.machines.tile.ItemStackToItemStackTileEntity;
 
 @BossToolsModElements.ModElement.Tag
 public class CompressorGuiGui extends BossToolsModElements.ModElement {
