+++ src/main/java/net/mrscauthd/boss_tools/machines/CompressorBlock.java
@@ -42,9 +42,9 @@ import net.mrscauthd.boss_tools.ModInnet;
 import net.mrscauthd.boss_tools.crafting.BossToolsRecipeTypes;
 import net.mrscauthd.boss_tools.crafting.ItemStackToItemStackRecipeType;
 import net.mrscauthd.boss_tools.gui.CompressorGuiGui;
-import net.mrscauthd.boss_tools.machines.machinetileentities.ItemStackToItemStackTileEntity;
-import net.mrscauthd.boss_tools.machines.machinetileentities.PowerSystem;
-import net.mrscauthd.boss_tools.machines.machinetileentities.PowerSystemCommonEnergy;
+import net.mrscauthd.boss_tools.machines.tile.ItemStackToItemStackTileEntity;
+import net.mrscauthd.boss_tools.machines.tile.PowerSystem;
+import net.mrscauthd.boss_tools.machines.tile.PowerSystemCommonEnergy;
 
 public class CompressorBlock {
 	public static class CustomBlock extends Block {
