+++ src/main/java/net/mrscauthd/boss_tools/machines/SolarPanelBlock.java
@@ -52,9 +52,9 @@ import net.minecraftforge.registries.ObjectHolder;
 import net.mrscauthd.boss_tools.BossToolsModElements;
 import net.mrscauthd.boss_tools.gui.SolarPanelGUIGui;
 import net.mrscauthd.boss_tools.itemgroup.BossToolsItemGroups;
-import net.mrscauthd.boss_tools.machines.machinetileentities.GeneratorTileEntity;
-import net.mrscauthd.boss_tools.machines.machinetileentities.PowerSystem;
-import net.mrscauthd.boss_tools.machines.machinetileentities.PowerSystemNone;
+import net.mrscauthd.boss_tools.machines.tile.GeneratorTileEntity;
+import net.mrscauthd.boss_tools.machines.tile.PowerSystem;
+import net.mrscauthd.boss_tools.machines.tile.PowerSystemNone;
 
 @BossToolsModElements.ModElement.Tag
 public class SolarPanelBlock extends BossToolsModElements.ModElement {
