+++ src/main/java/net/mrscauthd/boss_tools/machines/tile/GeneratorTileEntity.java
@@ -1,4 +1,4 @@
-package net.mrscauthd.boss_tools.machines.machinetileentities;
+package net.mrscauthd.boss_tools.machines.tile;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,6 +12,7 @@ import net.minecraftforge.common.capabilities.Capability;
 import net.minecraftforge.common.util.LazyOptional;
 import net.minecraftforge.energy.CapabilityEnergy;
 import net.minecraftforge.energy.IEnergyStorage;
+import net.mrscauthd.boss_tools.capability.EnergyStorageBasic;
 
 public abstract class GeneratorTileEntity extends AbstractMachineTileEntity {
