+++ src/main/java/net/mrscauthd/boss_tools/machines/tile/PowerSystemFuelAbstract.java
@@ -1,4 +1,4 @@
-package net.mrscauthd.boss_tools.machines.machinetileentities;
+package net.mrscauthd.boss_tools.machines.tile;
 
 import java.util.List;
