+++ src/main/java/net/mrscauthd/boss_tools/machines/tile/PowerSystemFuelGeneratingRecipe.java
@@ -1,4 +1,4 @@
-package net.mrscauthd.boss_tools.machines.machinetileentities;
+package net.mrscauthd.boss_tools.machines.tile;
 
 import net.minecraft.inventory.Inventory;
 import net.minecraft.item.ItemStack;
